4/exercise4_material/src_to_implement/requirements.txt
numpy
scipy
pandas
matplotlib
pytorch
torchvision
cuda100
pip
scikit-image
scikit-learn
jupyterlab
tqdm